aaold-old-code-base/requirements/dev.txt
# Development requirements
-r base.txt

# Django extensions for development utilities
django-extensions>=3.2.3

# Debug toolbar
django-debug-toolbar>=4.2.0

# Code formatting
black>=24.1.1
isort>=5.13.2

# Linting
flake8>=7.0.0
pylint>=3.0.3
pylint-django>=2.5.5

# Testing
pytest>=7.4.4
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Faker for test data
Faker>=22.2.0

# IPython for better shell
ipython>=8.20.0

# Watchdog for auto-reload
watchdog>=4.0.0
